web_viewer/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.1.5
